+++ R/dlm.main.R
@@ -3,6 +3,12 @@ dlm.main <- function(formula , data , x , y , q , remove , type = 1){
   remove.original = remove
   if (type == 1){
     remove = unlist(remove)
+    remove.intercept <- FALSE
+    if ( sum(remove %in% -1) >= 1 ){ # See if remove intercept
+      remove.intercept <- TRUE
+      remove <- remove[which(remove != -1)]
+    }
+    
     if (sum(remove == 0) > 0){ 
       remove.main = TRUE
       remove = remove[which(remove != 0)]
@@ -15,7 +21,13 @@ dlm.main <- function(formula , data , x , y , q , remove , type = 1){
     design[,2] = x[(q+1):n]
     design.colnames[1] = "y.t"
     design.colnames[2] = "x.t"
-    modelStr = "y.t ~ x.t +"
+    
+    if (remove.intercept){
+      modelStr = "y.t ~ -1 + x.t +"
+    } else {
+      modelStr = "y.t ~ x.t +"
+    }
+    
     seq = 3:(q+2) 
     seq = seq[ ! seq %in% (remove+2)] # remove the lags to be dropped from the list
     count = 2 
@@ -33,7 +45,13 @@ dlm.main <- function(formula , data , x , y , q , remove , type = 1){
       design <- design[-2]
     }
     
-    model = lm(y.t ~ . , design )
+    if (remove.intercept){
+      model.formula <- as.formula(y.t ~ . -1)
+    } else {
+      model.formula <- as.formula(y.t ~ .)
+    }
+
+    model = lm(formula = model.formula , design )
     
     output = list(model = model, designMatrix = design , q = q , removed = remove)
     model$call = "Y ~ X"
@@ -41,6 +59,12 @@ dlm.main <- function(formula , data , x , y , q , remove , type = 1){
     
     n=nrow(data)
     vars = formula.tools::get.vars(formula)
+    
+    remove.intercept <- FALSE
+    
+    if ( attr(terms(formula),"intercept") == 0 ){ # See if remove intercept
+      remove.intercept <- TRUE
+    }
     
     dep = vars[1] #get the name of dependent variable as a string
     indeps = vars[2:length(vars)] # get the names of independents variables
@@ -132,7 +156,15 @@ dlm.main <- function(formula , data , x , y , q , remove , type = 1){
     design.colnames = design.colnames[which(!is.na(design.colnames))]
     design = data.frame(depVar , design)
     colnames(design) = c(dep , design.colnames[1:(length(design)-1)])
-    model = lm(formula(substr(modelStr, 1, nchar(modelStr)-3) ) , design )
+    
+    model.formula <- substr(modelStr, 1, nchar(modelStr)-3) 
+   
+    if (remove.intercept){
+      model.formula <- paste0(model.formula, " -1")
+      model = lm(as.formula(model.formula) , design )
+    } else {
+      model = lm(as.formula(model.formula) , design )
+    }
     output = list(model = model, designMatrix = design , k = (length(vars) - 1) , q = q , removed = remove.original , formula = formula , data = data)
     model$call = toString(formula)
   }